Timeouts, Dialback cleanup
New outbound sessions now timeout queued stanzas, allowing them to be returned to the user. This is
triggered a configurable amount after the first stanza is enqueued, 10 seconds by default.

Dialback was also rewritten to avoid lots of very similar parsing code, which all looked very similar and
there was lots of it. This removed lots of very similar parsing code.

A couple of cases of dialback error handling were introduced.