Add Makefile rules to easily regenerate DH params
As @SimonWaters noted in #1, the use of hardcoded DH parameters is
undesirable. This doesn't change this state of affairs, but
mitigates it somewhat by making it simple to regenerate the
parameters during build, and provides both Makefile rules and
documentation to support this.

Users in sensitive locations are advised to make use of these.