Add XMPP version flag to domain configuration
The commit introduces a new flag, 'xmpp_ver', to the domain configuration which governs whether the XMPP version should be ignored in the stream open. This configuration can be modified per domain, allowing granular control over the handling of the XMPP version. The log will also include messages when the version is suppressed due to configuration.